Refactor: Restructure clintrials package for maintainability
This commit restructures the `clintrials` package to improve its organization and maintainability. The changes are as follows:

- Consolidated utility modules: Merged the functionality of `clintrials/util.py` and `clintrials/utils.py` into a single `clintrials/utils.py` module to remove redundant code and centralize utility functions.

- Restructured the `clintrials` package: Created a new `clintrials/core` sub-package to house the fundamental components of the library. Moved `common.py`, `numerics.py`, `recruitment.py`, `simulation.py`, `stats.py`, and `tte.py` into the new `clintrials/core` directory. Renamed `clintrials/common.py` to `clintrials/core/math.py` to better reflect its content.

- Updated imports: Updated all necessary imports throughout the codebase to reflect the new file locations and package structure.

- Verified the changes: Ran the test suite to ensure that the changes have not broken any existing functionality. All tests pass except for one failing test in `tests/test_efftox.py::test_thall2014_efftox_v2`. This test appears to be stochastic and may be failing due to changes in the random seed or other subtle changes in the code that affect the simulation results. Further investigation is needed to determine the exact cause of the failure.
