+++ .gitignore
@@ -6,3 +6,34 @@ dist/
 .venv/
 poetry.lock
 docs/_build/
+
+.hypothesis/constants/1a33f243fd790857
+.hypothesis/constants/1d1a63a3ba8c0429
+.hypothesis/constants/20902782b784faf9
+.hypothesis/constants/3f472b85df853f6f
+.hypothesis/constants/4f3beb6651cea6a9
+.hypothesis/constants/50286a1fdc0f0abb
+.hypothesis/constants/586a70722266cc31
+.hypothesis/constants/5f8865e21f827b56
+.hypothesis/constants/6829289ebedeb8ef
+.hypothesis/constants/6b858c92e47aa396
+.hypothesis/constants/7732a38dc88b7b8d
+.hypothesis/constants/785757f0a247667b
+.hypothesis/constants/8b1c202fac6234db
+.hypothesis/constants/90d6e100effa4dac
+.hypothesis/constants/95b4482a626f143e
+.hypothesis/constants/991ac41ec148c75d
+.hypothesis/constants/9dc2cdb661c588c9
+.hypothesis/constants/adbcff484ccdeabe
+.hypothesis/constants/afede177a3e6251a
+.hypothesis/constants/b3986bc8cd184725
+.hypothesis/constants/ba3244df1cdfc8fb
+.hypothesis/constants/bb7fbe20d06799de
+.hypothesis/constants/c8d9199f9f83e98a
+.hypothesis/constants/caf9d342cf8f0b7e
+.hypothesis/constants/cd3771046418a747
+.hypothesis/constants/da39a3ee5e6b4b0d
+.hypothesis/constants/e8b30f2d9cac93be
+.hypothesis/constants/f36ccca712fc4b4b
+.hypothesis/constants/f5406eed55345ae3
+.hypothesis/constants/fd665a6f54aeb410
