+++ .hypothesis/constants/116109af51dc2afd
@@ -1,0 +1,4 @@
+# file: /app/clintrials/dosefinding/crm.py
+# hypothesis_version: 6.138.2
+
+[0.0, 0.05, 0.95, 1.34, 2.0, 100, 200, 300, 1000000, '-', '--', '-.', 'BFGS', 'BetaHat', 'BetaVar', 'CRM', 'Dose level', 'Kristian Brock', 'Posterior', 'Prior', 'ProbTox', 'Type', 'Unknown backend', 'bayes', 'black', 'bootstrap', 'crm', 'crm_dtp_detail', 'hessian', 'k', 'laplace', 'logit1', 'mle', 'none', 'o', 'quadrature', 'x']
