+++ .hypothesis/constants/16966faf82c35e78
@@ -1,0 +1,4 @@
+# file: /app/clintrials/core/__init__.py
+# hypothesis_version: 6.138.2
+
+[]
