+++ .hypothesis/constants/43313d65d8481cbd
@@ -1,0 +1,4 @@
+# file: /app/clintrials/dashboard/views/efftox_view.py
+# hypothesis_version: 6.138.2
+
+[0.05, 0.1, 0.2, 0.3, 0.4, 0.5, 0.6, 'Dose Level', 'N', 'Probability', 'Probability Type', 'Simulation Summary', 'group', 'prob_accept_eff', 'prob_accept_tox', 'recommended_dose', 'true_prob_eff', 'true_prob_tox']
