+++ .hypothesis/constants/4d27998df49ce21e
@@ -1,0 +1,4 @@
+# file: /app/clintrials/dashboard/views/winratio_view.py
+# hypothesis_version: 6.138.2
+
+[0.0, 0.05, 0.25, 0.27, 0.43, 0.5, 0.75, 1.0, 100, 1000, 'Results', 'Run Simulation', 'Significance level', 'Simulation complete', 'Win Ratio Simulation']
