+++ .hypothesis/constants/9b2f98a0412ca4aa
@@ -1,0 +1,4 @@
+# file: /app/clintrials/core/math.py
+# hypothesis_version: 6.138.2
+
+['Kristian Brock']
