+++ .hypothesis/constants/a26964670c6899ee
@@ -1,0 +1,4 @@
+# file: /app/clintrials/dosefinding/watu.py
+# hypothesis_version: 6.138.2
+
+[0.05, 1.34, 'Kristian Brock']
