+++ .hypothesis/constants/c4e9b64168cc84bc
@@ -1,0 +1,4 @@
+# file: /app/clintrials/dosefinding/wagestait.py
+# hypothesis_version: 6.138.2
+
+[0.025, 1.34, 100, 'Kristian Brock']
