+++ .hypothesis/constants/f1d4785a00a92ee1
@@ -1,0 +1,4 @@
+# file: /app/clintrials/dashboard/views/crm_view.py
+# hypothesis_version: 6.138.2
+
+[0.05, 0.1, 0.2, 0.3, 0.4, 0.5, 'Dose Level', 'N', 'Probability', 'Simulation Summary', 'group', 'recommended_dose', 'true_tox']
