+++ clintrials/core/__init__.py
@@ -1,0 +1,1 @@
+# This file marks the `core` directory as a Python package.
