+++ clintrials/core/numerics.py
