+++ clintrials/dosefinding/efftox.py
@@ -16,10 +16,10 @@ from collections import OrderedDict
 import numpy as np
 from scipy.optimize import brentq
 
-from clintrials.common import inverse_logit
+from clintrials.core.math import inverse_logit
+from clintrials.core.stats import ProbabilityDensitySample
 from clintrials.dosefinding.efficacytoxicity import EfficacyToxicityDoseFindingTrial
-from clintrials.stats import ProbabilityDensitySample
-from clintrials.util import atomic_to_json, iterable_to_json
+from clintrials.utils import atomic_to_json, iterable_to_json
 
 
 def scale_doses(real_doses):
