+++ tests/test_logging.py
@@ -1,8 +1,8 @@
 import logging
 
+from clintrials.core.simulation import run_sims
 from clintrials.dosefinding import find_mtd
 from clintrials.logging import get_logger
-from clintrials.simulation import run_sims
 
 
 def dummy_sim():
