+++ tests/test_recruitment.py
@@ -5,7 +5,7 @@ import numpy as np
 import pytest
 from numpy.testing import assert_almost_equal
 
-from clintrials.recruitment import (
+from clintrials.core.recruitment import (
     ConstantRecruitmentStream,
     QuadrilateralRecruitmentStream,
 )
